refactor: move peer message policy into module
