+++ crates/fiber-lib/src/fiber/mod.rs
@@ -16,6 +16,7 @@ mod key;
 #[cfg(not(target_arch = "wasm32"))]
 pub(crate) mod onion_service;
 mod path;
+mod peer_message_policy;
 #[cfg(not(target_arch = "wasm32"))]
 pub(crate) mod proxy;
 mod settle_tlc_set_command;
